requirements.txt
redis
Flask
gunicorn==20.1.0
InstructorEmbedding
python-dotenv
sentence_transformers